RENAME: convergence_analysis_models -> analysis_models
