+++ README.md
@@ -42,7 +42,7 @@ homogeneous Dirichlet conditions on a 3D simplex grid:
   * [runscript_time_convergence_dirichlet_boundaries](./convergence_analysis/runscript_time_convergence_dirichlet_boundaries.py) 
 
 All the runscripts utilize
-[manufactured_solution_dynamic_3D](./convergence_analysis/convergence_analysis_models/manufactured_solution_dynamic_3D.py)
+[manufactured_solution_dynamic_3D](./convergence_analysis/analysis_models/manufactured_solution_dynamic_3D.py)
 as the manufactured solution setup.
 
 ### Convergence analysis of MPSA-Newmark with absorbing boundaries
@@ -52,13 +52,13 @@ boundaries have roller conditions, and the left boundary is a time-dependent Dir
 condition which sends in the orthogonal wave.
 
 Convergence in space and time, isotropic:
-  * [runscript_space_time_convergence_absorbing_boundaries](./convergence_analysis/runscript_space_time_convergence_absorbing_boundaries.py) which uses the model class setup found in [model_convergence_ABC](./convergence_analysis/convergence_analysis_models/model_convergence_ABC.py)
+  * [runscript_space_time_convergence_absorbing_boundaries](./convergence_analysis/runscript_space_time_convergence_absorbing_boundaries.py) which uses the model class setup found in [model_convergence_ABC](./convergence_analysis/analysis_models/model_convergence_ABC.py)
 
 Convergence in space and time, vertical strip inside the domain which is anisotropic in
 directions orthogonal to the propagating wave:
   * [runscript_space_time_absorbing_boundaries_anisotropy](./convergence_analysis/runscript_space_time_absorbing_boundaries_anisotropy.py)
     which uses the model class setup found in
-    [model_convergence_ABC_anisotropy](./convergence_analysis/convergence_analysis_models/model_convergence_ABC_anisotropy.py)
+    [model_convergence_ABC_anisotropy](./convergence_analysis/analysis_models/model_convergence_ABC_anisotropy.py)
 
 
 ### Energy decay analysis of MPSA-Newmark with absorbing boundaries
