+++ convergence_analysis/analysis_models/model_energy_decay_analysis.py
