+++ convergence_analysis/runscript_space_convergence_dirichlet_boundaries.py
@@ -13,7 +13,7 @@ import porepy as pp
 from porepy.applications.convergence_analysis import ConvergenceAnalysis
 
 sys.path.append("../")
-from convergence_analysis_models.manufactured_solution_dynamic_3D import ManuMechSetup3d
+from analysis_models.manufactured_solution_dynamic_3D import ManuMechSetup3d
 from utils_convergence_analysis import export_errors_to_txt, run_analysis
 
 # Prepare path for generated output files
