+++ convergence_analysis/runscript_space_time_convergence_dirichlet_boundaries.py
@@ -17,7 +17,7 @@ sys.path.append("../")
 from plotting.plot_utils import draw_multiple_loglog_slopes, fetch_numbers_from_file
 from porepy.applications.convergence_analysis import ConvergenceAnalysis
 
-from convergence_analysis_models.manufactured_solution_dynamic_3D import ManuMechSetup3d
+from analysis_models.manufactured_solution_dynamic_3D import ManuMechSetup3d
 from utils_convergence_analysis import export_errors_to_txt, run_analysis
 
 # Prepare path for generated output files
